Moved StreamForwarder to public utils package
Also added unit test for it
